requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
efficientnet==1.1.1
entrypoints==0.3
gast==0.3.3
google-pasta==0.2.0
graphviz==0.14.1
grpcio==1.29.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.6.1
imutils==0.5.3
install==1.3.3
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
llvmlite==0.34.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numba==0.51.2
numpy==1.18.5
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.0
PCV==1.0
pickleshare==0.7.5
Pillow==7.2.0
progressbar==2.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
pydot==1.4.1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.24.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.7.24
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
urllib3==1.25.10
utility==1.0
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
